+++ ComicsShelf.UWP/Helpers/FileSystem_Permissions.cs
@@ -1,0 +1,12 @@
+﻿namespace ComicsShelf.UWP
+{
+   partial class FileSystem
+   {
+
+      public void CheckPermissions(System.Action grantedCallback, System.Action revokedCallback)
+      {
+         grantedCallback?.Invoke();
+      }
+
+   }
+}
