+++ ComicsShelf/Helpers/Dependency/iFileSystem.cs
@@ -1,25 +1,34 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Xamarin.Forms;
 
 namespace ComicsShelf.Helpers
 {
-   public interface iFileSystem
+
+   public interface iFileSystem : IDisposable
    {
+
       string PathSeparator { get; }
-      Task<string> GetCachePath();
-      Task<string> GetDataPath();
+
+      string GetCachePath();
+      string GetDataPath();
+
       Task<bool> ValidateLibraryPath(string libraryPath);
       Task<string> GetLibraryPath();
+
       Task<string[]> GetFiles(string path);
-      Task<System.IO.Compression.ZipArchive> GetZipArchive(Settings.Settings settings, File.FileData comicFile);
-      Task<bool> FileExists(string path);
+
+      Task<System.IO.Compression.ZipArchive> GetZipArchive(Helpers.Settings.Settings settings, string fullPath);
       Task Thumbnail(System.IO.Stream imageStream, string imagePath);
 
-      void CheckPermissions(System.Action grantedCallback, System.Action revokedCallback);
+      void CheckPermissions(Action grantedCallback, Action revokedCallback);
    }
+
    public class FileSystem
    {
       public static iFileSystem Get()
       { return DependencyService.Get<iFileSystem>(); }
+
    }
+
 }
